+++ docs/15_0_0/gettingstarted/whatsnew.md
@@ -9,6 +9,7 @@ This page contains a list of big features. Please check the GitHub issues for al
 Look into [migration guide](https://primefaces.github.io/primefaces/15_0_0/#/../migrationguide/15_0_0) for more enhancements and changes.
 
 * Core
+    * ???
     
 * AutoComplete
     * Added property `highlightSelector=""` so you can delcare the jQuery selector for what to find and highlight. See https://github.com/primefaces/primefaces/issues/11822 
@@ -18,3 +19,6 @@ Look into [migration guide](https://primefaces.github.io/primefaces/15_0_0/#/../
     
 * FeedReader
     * Added `podcast="true"` property if [Apple Itunes Podcast](https://help.apple.com/itc/podcasts_connect/#/itcb54353390) parsing and specific tags 
+    
+* OutputLabel
+    * Added `disabled="true"` property to allow OutputLabel to be disabled like other components